Puntos f y h terminados. Se agrego el archivo json de la coleccion de endpoints de postman

+++ integrador3/src/main/java/com/tp/integrador3/service/dto/estudiantecarrera/response/EstudianteCarreraResponseDTO.java
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class EstudianteCarreraResponseDTO {
 
-    private final Integer id;
+    private final long id;
     private final String estudianteNombre;
     private final String estudianteApellido;
     private final String carreraNombre;
